windows fixes

+++ include/nori/kdtree.h
@@ -499,7 +499,7 @@ protected:
         );
 
         NodeType &splitNode = m_nodes[*split];
-        splitNode.setAxis(axis);
+        splitNode.setAxis((uint8_t) axis);
         splitNode.setLeaf(false);
 
         if (split+1 != rangeEnd)
